Read OHLC CSV files using Google's "data.table" library.
"data.table" is a replacement for "data.frame" that is more
flexible. More importanly for my purposes it is faster, especially when it
comes to reading CSV files (I have tested at 1 to 2 orders of magnitude
faster.) data.table is also worth learning because it is implemented as a
derived class of data.frame. So any library that takes a data.frame object
(which is all financially libraries practically), can take a data.table
object.  This will be especially helpful when running large ML models time and
again where data loading can take a signifcant portion of the time.  Finally,
data.table brings a new paradigm to R for shaping and filtering data which is
very powerful.