Load sqlite 5 sec OHLC data & return 1 min xts object.
This is a complex routine due to lock contention when sqlite has
concurrent reads and writes.  Since I want to support multiple processes
accessing the sqlite DB I have to handle the conflict here.  Of course, MySql,
Postgres, or any other server based RDBMS would handle this automatically, but
that's the trade off for using SQLite